django-pjt/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.1.5
django-seed==0.2.2
djangorestframework==3.15.2
Faker==33.0.0
idna==3.10
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.10.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
